Some audio file stuff and fixes for builds having added new files.

+++ src/plank/files/plank_IffFileCommon.h
@@ -151,7 +151,7 @@ static inline PlankResult pl_IffFile_InitID (PlankIffFileRef p, const char* stri
     }
     else
     {
-        length = strlen (string);
+        length = (int)strlen (string);
         
         switch (length)
         {
